Split WpcomChecklist into a loader and a loaded component.
WpcomChecklist loads a sizeable amount of code, but it's often only
needed in sites that qualify and haven't completed the checklist.

Instead of statically including and instancing the component every
time, this PR instead extracts some minimal exports to determine
whether it makes sense to display the WpcomChecklist content. Only
if these checks are passed does WpcomChecklistComponent then get
asynchronously loaded and instanced.
